refactor: decoupling demos from advanced_sorting_algorithms
